fixed more warnings for the sun studio compiler and fixed parse time type checking for the + and - operators
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/trunk@3907 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ lib/QC_Socket.cpp
@@ -514,11 +514,11 @@ static AbstractQoreNode *SOCKET_setCharset(QoreObject *self, mySocket *s, const 
 
 // Socket::isDataAvailable(timeout $timeout_ms = 0)
 static AbstractQoreNode *SOCKET_isDataAvailable(QoreObject *self, mySocket *s, const QoreListNode *params, ExceptionSink *xsink) {
-   return get_bool_node(s->isDataAvailable(HARD_QORE_INT(params, 0)));
+   return get_bool_node(s->isDataAvailable((int)HARD_QORE_INT(params, 0)));
 }
 
 static AbstractQoreNode *SOCKET_isWriteFinished(QoreObject *self, mySocket *s, const QoreListNode *params, ExceptionSink *xsink) {
-   return get_bool_node(s->isWriteFinished(HARD_QORE_INT(params, 0)));
+   return get_bool_node(s->isWriteFinished((int)HARD_QORE_INT(params, 0)));
 }
 
 static AbstractQoreNode *SOCKET_isOpen(QoreObject *self, mySocket *s, const QoreListNode *params, ExceptionSink *xsink) {
